Set up current motor and encoder testing for open loop parameterization
need to ask ridgely about how to write queues with vectors. This just puts the position, velocity, and time in a queue and then prints them out so the data can be copied to a textfile and then sent to excel, might need to tune timing on encoder task to get good instanteneous velocity measurements.