Hardware test tasks implemented
Two new testing sections in the code exist in main; a constant velocity test and a control path test. These can be used to test the hardware of the system.

Additionally, commands.cpp and commands.h have been obsoleted, and the functions that were once in commands have been transferred to translate.cpp and translate.h.